+++ src/Exceptional.php
@@ -12,6 +12,7 @@ namespace DecodeLabs;
 use BadMethodCallException;
 use DecodeLabs\Exceptional\Exception;
 use DecodeLabs\Exceptional\Factory;
+use Throwable;
 
 final class Exceptional
 {
@@ -20,6 +21,43 @@ final class Exceptional
      */
     private function __construct()
     {
+    }
+
+    /**
+     * Generic call signature for PHPStan
+     *
+     * @param array<string,mixed> $params
+     * @param array<string> $interfaces
+     * @param array<string> $traits
+     */
+    private static function _phpstan(
+        ?string $message = null,
+        ?array $params = [],
+        mixed $data = null,
+        ?Throwable $previous = null,
+        ?int $code = null,
+        ?string $file = null,
+        ?int $line = null,
+        ?int $http = null,
+        ?string $namespace = null,
+        ?array $interfaces = null,
+        ?array $traits = null
+    ): Exception {
+        return Factory::create(
+            ['Generic'],
+            1,
+            $message,
+            $params,
+            $data,
+            $previous,
+            $code,
+            $file,
+            $line,
+            $http,
+            $namespace,
+            $interfaces,
+            $traits
+        );
     }
 
     /**
@@ -52,7 +90,6 @@ final class Exceptional
         return Factory::create(
             explode(',', $type),
             1,
-            /* @phpstan-ignore-next-line */
             ...$args
         );
     }